requirements.txt
pandas>=1.0.0
python-dotenv>=0.15.0
requests>=2.0.0
openpyxl>=3.0.0
